- A new reusable function  "showjsonError" is created, to use it in all   other functions. (Under package "utils". - Setting the functions (getAllMovies & getMovieWithSpecificID) to a   private functions, since i'm handling them in 'HandleGetMovies', and   no need to access them in the main functions. - I've created a Postman Collection, including all methods.
