+++ .vscode/launch.json
@@ -1,11 +1,11 @@
 {
     "configurations": [
         {
-            "name": "Launch .moviesUsersAPI/main.go",
+            "name": "Launch main.go",
             "type": "go",
             "request": "launch",
             "mode": "debug",
-            "program": "./moviesUsersAPI/main.go"
+            "program": "main.go"
         }
     ]
 }
