+++ README.md
@@ -1,0 +1,30 @@
+# Movie Metadata API
+
+This is a simple API written in Go that handles basic movie metadata. It includes:
+
+- Adding new movies
+- Fetching all movies
+- Getting movie details by ID
+- Tracking when a user watches a movie (stored in memory)
+
+## Running the Project
+
+Make sure Go is installed, then run:
+
+```bash
+go run main.go
+
+
+####### 📥 Postman Collection
+
+Kindly, You can access and test all API endpoints using this shared Postman collection:
+
+👉 [Open Postman Collection](
+    https://issa-3986679.postman.co/workspace/Issa%27s-Workspace~cc04252e-3329-4cf9-8c9f-ceef5268c2a8/collection/46421072-9c0f73e8-2109-47ec-abe5-2bc38b666ea5?action=share&creator=46421072
+    
+    )
+
+To use:
+1. Click the link above
+2. Open it in Postman (or fork it to your workspace)
+3. Use the requests directly or modify as needed
